menambahkan function untuk menampilkan data, un/verfy data, delete, dan validasinya

+++ app/Controllers/Admin.php
@@ -1,7 +1,6 @@
 <?php
 
 namespace App\Controllers;
-use App\Models\Users_model;  //mengambil Model pada folder Model dengan menggunakan namespace
 use App\Models\Admin_model;  
 use App\Models\Post_model;  
 use App\Models\Games_model;  
@@ -17,7 +16,6 @@ class Admin extends BaseController {
 	protected $validation;
 
 	public function __construct() {
-		$this->users_model = new Users_model(); //untuk memanggil model sekali dan bisa digunakan berkali2
 		$this->admin_model = new Admin_model();
 		$this->post_model = new Post_model();
 		$this->games_model = new Games_model();
@@ -32,66 +30,50 @@ class Admin extends BaseController {
 
 	//ini buat halaman dashboard admin
 	public function index() {
-		$where_condition = array('p.verified' => false);
-		$post_data = $this->post_model->get_post($where_condition)->getResultArray(); //post data hanya menampilkan image, score, dan username
-        // dd($this->post_model->coba()->getResultArray());
-        
-        
-		
-		// if ($this->session->get('logged_in') && $this->session->get('level') == 'admin') {
-        // }
-        
-		// return redirect()->to('/login');
-        $data = [
-            'title' => 'Admin', //sebagai judul page, dikirim ke <title> pada view
-            'session_data' => $this->session->get(),
-            'posts' => $post_data,
-            'games' => $this->games
-        ];
-        return view('admin/indexAdmin', $data);	
-	}
-	
-	//buat nampilin halaman home buat admin, untuk mengecek apakah sudah bekerja dari otoritas adminnya
-	public function home() {
-		$where_condition = ['p.verified' => true];
-		$post_data = $this->post_model->get_post($where_condition); //post data hanya menampilkan image, score, dan username
-		
-		// if ($this->session->get('logged_in') && $this->session->get('level') == 'admin') {
-				
-		// }
-
-		// return redirect()->to('/login');
-        $data = [
+		$keyword = $this->req->getVar('keyword');
+		if ($keyword) {
+			$post_data = $this->admin_model->search_InAdmin($keyword)->getResultArray();
+		} else {
+			$post_data = $this->admin_model->get_post_Inadmin()->getResultArray(); //post data hanya menampilkan image, score, dan username
+		}
+    
+		if ($this->session->get('logged_in') && $this->session->get('level') == 'admin') {
+			$data = [
 				'title' => 'Admin', //sebagai judul page, dikirim ke <title> pada view
 				'session_data' => $this->session->get(),
 				'posts' => $post_data,
 				'games' => $this->games
 			];
-			return view('admin/homeAdmin', $data);
-	
+			return view('admin/indexAdmin', $data);	
+			
+        }
+        
+		return redirect()->to('/login');
 	}
-
-	//buat nampilin halaman admin
 	
-
 	public function games($id) {
 		$where_condition = [
-            'p.verified' => true,
             'p.game_id' => $id
         ];
-		$post_data = $this->post_model->get_post($where_condition); //post data hanya menampilkan image, score, dan username
+		$keyword = $this->req->getVar('keyword');
+
+		if ($keyword) {
+			$post_data = $this->admin_model->search_InAdmin($keyword)->getResultArray();
+		} else {
+			$post_data = $this->post_model->get_post($where_condition)->getResultArray(); //post data hanya menampilkan image, score, dan username
+		}
 		
-		// if ($this->session->get('logged_in') && $this->session->get('level') == 'admin') {
-        // }
+		if ($this->session->get('logged_in') && $this->session->get('level') == 'admin') {
+			$data = [
+				'title' => 'Admin', //sebagai judul page, dikirim ke <title> pada view
+				'session_data' => $this->session->get(),
+				'posts' => $post_data,
+				'games' => $this->games
+			];
+			return view('admin/gamesAdmin', $data);	
+        }
         
-		// return redirect()->to('/login');
-        $data = [
-            'title' => 'Admin', //sebagai judul page, dikirim ke <title> pada view
-            'session_data' => $this->session->get(),
-            'posts' => $post_data,
-            'games' => $this->games
-        ];
-        return view('admin/gamesAdmin', $data);	
+		return redirect()->to('/login');
 	}
 
 	public function logout() {
@@ -102,12 +84,32 @@ class Admin extends BaseController {
 	}
 
 	public function verify($id) {
-        $this->post_model->verify($id);
-        return redirect()->to('/adm');
+		if ($this->session->get('logged_in') && $this->session->get('level') == 'admin') {
+			$this->post_model->verify($id);
+			return redirect()->to('/admin');
+		}
+
+		return redirect()->to('/login');
+	}
+
+	public function unverify($id) {
+		if ($this->session->get('logged_in') && $this->session->get('level') == 'admin') {
+			$this->post_model->verify($id);
+			return redirect()->to('/admin');
+		}
+
+		return redirect()->to('/login');
 	}
 
 	public function delete($id) {
-		$this->post_model->delete($id);
-		return redirect()->to('/adm');
+		if ($this->session->get('logged_in') && $this->session->get('level') == 'admin') {
+			$this->post_model->delete($id);
+			return redirect()->to('/admin');
+		}
+
+		return redirect()->to('/login');
 	}
+
+	
+
 }
